requirements.txt
tqdm
docx2txt
python-docx
regex